Fix splitting data when there are negatve values
