+++ packages/perspective-viewer-d3fc/src/js/axis/mainAxis.js
@@ -12,15 +12,18 @@ import * as fc from "d3fc";
 export const scale = settings => d3.scaleLinear();
 
 export const domain = (settings, data) => {
+    // For multiple main values, we need to check the maximum and minimum (+ve and -ve)
+    const mainValueAccessors = aggFn => settings.mainValues.map((m, i) => (d => d.reduce((max, v) => aggFn(max, v.mainValue), 0)));
+
     const accessors = (settings.mainValues.length > 1)
-        ? settings.mainValues.map((m, i) => (d => d.reduce((max, v) => Math.max(max, v.mainValue), 0)))
+        ? mainValueAccessors(Math.max).concat(mainValueAccessors(Math.min))
         : [d => d.mainValue];
     
     return fc.extentLinear()
         .include([0])
         .pad([0, 0.1])
         .accessors(accessors)
-        (data[data.length -1]);
+        (data.flat());
     }
 
 export const label = settings => settings.mainValues.map(v => v.name).join(', ');
